chore: add unit tests for the game
