+++ go/game.go
@@ -1,6 +1,10 @@
 package main
 
-import "fmt"
+import (
+	"fmt"
+	"os"
+	"strings"
+)
 
 type Television interface {
 	Show(message string)
@@ -44,4 +48,33 @@ func (g Game) Play() {
 	}
 
 	tv.Show("🏆 Hero " + hero.Name + " wins!")
+}
+
+func loadGame(f string) Game {
+	// load game from file
+	content, err := os.ReadFile(f)
+	if err != nil {
+		panic(err)
+	}
+
+	// convert content into array of string lines
+	lines := strings.Split(string(content), "\n")
+
+	// first line is the hero
+	hero := NewHero(lines[0])
+
+	rooms := []Room{}
+	for _, line := range lines[1:] {
+		// skip empty lines
+		if line == "" {
+			continue
+		}
+
+		rooms = append(rooms, NewRoom(line))
+	}
+
+	return Game{
+		hero: hero,
+		maze: rooms,
+	}
 }
