+++ go/main.go
@@ -34,32 +34,3 @@ func main() {
 
 	game.Play()
 }
-
-func loadGame(f string) Game {
-	// load game from file
-	content, err := os.ReadFile(f)
-	if err != nil {
-		panic(err)
-	}
-
-	// convert content into array of string lines
-	lines := strings.Split(string(content), "\n")
-
-	// first line is the hero
-	hero := NewHero(lines[0])
-
-	rooms := []Room{}
-	for _, line := range lines[1:] {
-		// skip empty lines
-		if line == "" {
-			continue
-		}
-
-		rooms = append(rooms, NewRoom(line))
-	}
-
-	return Game{
-		hero: hero,
-		maze: rooms,
-	}
-}
